# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AR Demo and candidate API
  version: 0.2.0
  description: 'Demo: RuoYi + PostgreSQL, fixed-role administrators and scenes. Docker
    Compose and local ArtifactStorage infrastructure are available without artifact
    HTTP endpoints. Historical cloud/day-night resource and publication contracts
    below await a real deliverable and client loading agreement; they are not implemented.'
servers:
- url: https://api.example.invalid
  description: placeholder
paths:
  /api/v1/scenes:
    post:
      operationId: createScene
      summary: 创建场景（包含地点信息）
      description: 创建场景（包含地点信息）
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneView'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneCreate'
      tags:
      - Demo 已实现
    get:
      operationId: listScenes
      summary: 列出场景
      description: 列出场景
      security:
      - bearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: name
        in: query
        required: false
        schema:
          type: string
          maxLength: 200
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneList'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-implementation: implemented
      tags:
      - Demo 已实现
  /api/v1/scenes/{sceneId}:
    get:
      operationId: getScene
      summary: 查看场景和当前昼夜版本
      description: 查看场景和当前昼夜版本
      security:
      - bearerAuth: []
      parameters:
      - name: sceneId
        in: path
        required: true
        schema: &id001
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneView'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-implementation: implemented
      tags:
      - Demo 已实现
    put:
      operationId: updateScene
      summary: 修改场景基本信息
      description: 修改场景基本信息
      security:
      - bearerAuth: []
      parameters:
      - name: sceneId
        in: path
        required: true
        schema: *id001
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneView'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneUpdate'
      tags:
      - Demo 已实现
  /api/v1/scenes/{sceneId}/enabled:
    put:
      operationId: setSceneEnabled
      summary: 启用或禁用场景
      description: 启用或禁用场景
      security:
      - bearerAuth: []
      parameters:
      - name: sceneId
        in: path
        required: true
        schema: *id001
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneView'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-implementation: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnabledUpdate'
      tags:
      - Demo 已实现
...
  /api/v1/audits:
    get:
      operationId: listAudits
      summary: 只读场景业务审计
      description: 只读场景业务审计
      security:
      - bearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditList'
        '400':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Request failed; 409 covers state/version conflict or incompatible
            client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-implementation: implemented
      tags:
      - Demo 已实现
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
      required:
      - code
      - message
      - traceId
...
    SceneCreate:
      type: object
      additionalProperties: false
      properties:
        name: &id005
          type: string
          minLength: 1
          maxLength: 200
        address: &id006
          type: string
          maxLength: 1000
          nullable: true
        longitude: &id007
          type: number
          minimum: -180
          maximum: 180
          nullable: true
        latitude: &id008
          type: number
          minimum: -90
          maximum: 90
          nullable: true
        geoCrs: &id009
          type: string
          enum:
          - WGS84
          - GCJ02
          - BD09
          nullable: true
      required:
      - name
    SceneView:
      type: object
      additionalProperties: false
      properties:
        name: *id005
        address: *id006
        longitude: *id007
        latitude: *id008
        geoCrs: *id009
        id: *id001
        enabled:
          type: boolean
        lockVersion:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - id
      - name
      - enabled
      - lockVersion
      - createdAt
      - updatedAt
    SceneList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SceneView'
        total:
          type: integer
          minimum: 0
      required:
      - items
      - total
...
    EnabledUpdate:
      type: object
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        expectedVersion:
          type: integer
          minimum: 0
        reason:
          type: string
          minLength: 1
      required:
      - enabled
      - expectedVersion
      - reason
...
    SceneUpdate:
      type: object
      additionalProperties: false
      properties:
        name: *id005
        address: *id006
        longitude: *id007
        latitude: *id008
        geoCrs: *id009
        expectedVersion:
          type: integer
          minimum: 0
      required:
      - name
      - expectedVersion
...
    AuditList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
          minimum: 0
      required:
      - items
      - total